Move withImage to legacy folder
