+++ react/legacy/FooterPaymentFormList.js
@@ -2,7 +2,7 @@ import React from 'react'
 import PropTypes from 'prop-types'
 
 import footerList from './footerList'
-import withImage from '../components/withImage'
+import withImage from './withImage'
 import footer from './footer.css'
 
 /**
@@ -14,17 +14,18 @@ const FooterPaymentFormItem = ({ imageSrc }) => {
   }
 
   return (
-    <img className={`${footer.paymentFormItem} pr0 w2 h2`} src={imageSrc} />
+    <img
+      className={`${footer.paymentFormItem} pr0 w2 h2`}
+      src={imageSrc}
+      alt=""
+    />
   )
 }
 
 FooterPaymentFormItem.displayName = 'FooterPaymentFormItem'
 
 FooterPaymentFormItem.propTypes = {
-  /** Indicates which one of the payments forms should the component show its image */
-  paymentType: PropTypes.string.isRequired,
-  /** If true, the original logo (with color) is used. If not, the grayscale's one */
-  showInColor: PropTypes.bool.isRequired,
+  imageSrc: PropTypes.string.isRequired,
 }
 
 const getImagePathFromProps = ({ paymentType, showInColor }) =>
