+++ react/legacy/withImage.d.ts
