+++ react/legacy/withImage.js
